feature: add temperature, frequency_penalty, presence_penalty, stop_tokens to generate_from_chunks request
